+++ spec/arachni/options_spec.rb
@@ -464,6 +464,37 @@ describe Arachni::Options do
         end
     end
 
+    describe '#exclude_body=' do
+        it 'should convert its param to an array of strings' do
+            exclude_body = %w(my_ignore my_other_ignore)
+
+            @opts.exclude_body = /test/
+            @opts.exclude_body.should == [/test/]
+
+            @opts.exclude_body = exclude_body.first
+            @opts.exclude_body.should == [Regexp.new( exclude_body.first )]
+
+            @opts.exclude_body = exclude_body
+            @opts.exclude_body.should == exclude_body.map { |p| Regexp.new( p ) }
+        end
+    end
+
+    describe '#exclude_body?' do
+        context 'when the string matches one of the #ignore patterns' do
+            it 'should return true' do
+                @opts.exclude_body = /test/
+                @opts.exclude_body?( 'this is a test test test' ).should be_true
+            end
+        end
+        context 'when the string does not match one of the #ignore patterns' do
+            it 'should return false' do
+                @opts.exclude_body = /test/
+                @opts.exclude_body?( 'this is a blah blah blah' ).should be_false
+            end
+        end
+    end
+
+
     describe '#lsmod=' do
         it 'should convert its param to an array of strings' do
             lsmod = %w(my_lsmod my_other_lsmod)